--- src/components/organisms/NewStartProjectPane/index.ts
@@ -1,3 +1,0 @@
-import NewStartProjectPane from './NewStartProjectPane';
-
-export {NewStartProjectPane};
