--- src/components/organisms/RecentProjectsPane/index.ts
@@ -1,1 +1,0 @@
-export {default} from './RecentProjectsPane';
