+++ src/components/organisms/StartProjectPane/NewRecentProjectsPane.tsx
@@ -1,0 +1,5 @@
+const NewRecentProjectsPane = () => {
+  return <p>Hello</p>;
+};
+
+export default NewRecentProjectsPane;
