+++ src/components/organisms/StartProjectPane/index.ts
@@ -1,1 +1,6 @@
-export {default} from './StartProjectPane';
+import NewRecentProjectsPane from './NewRecentProjectsPane';
+import NewStartProjectPane from './NewStartProjectPane';
+import RecentProjectsPane from './RecentProjectsPane';
+import StartProjectPane from './StartProjectPane';
+
+export {StartProjectPane, RecentProjectsPane, NewStartProjectPane, NewRecentProjectsPane};
